refactor: move around and add border
